requirements.txt
altgraph
contourpy
cycler
et_xmlfile
fonttools
kiwisolver
macholib
matplotlib
mpmath
mysql-connector
numpy
openpyxl
packaging
pandas
pillow
pyinstaller
pyinstaller-hooks-contrib
pyparsing
python-dateutil
pytz
PyYAML
setuptools
six
sp
sympy
tabulate
tzdata
xlrd
